Moved RunnableMessageReader and ExecutableMessageFactory into Connection class.
